fixed memleaks -- tutorial.cpp

+++ Project/Characters.cpp
@@ -890,7 +890,7 @@ namespace Characters
 				enemystats->EnemyType = NORMAL;
 				enemystats->health = 40 + (5 * enemystats->EnemyLevel);
 				enemystats->maxhealth = 40 + (5 * enemystats->EnemyLevel);
-				enemystats->enemytypedamage = 5 * enemystats->EnemyLevel;
+				enemystats->enemytypedamage = (f32)(5 * enemystats->EnemyLevel);
 				enemystats->damage = enemystats->enemytypedamage;
 				enemystats->EnemyCD = 2.0f;				//Cooldown till next enemy attack
 				enemystats->EnemyXP = 20 + (2 * enemystats->EnemyLevel);
@@ -901,7 +901,7 @@ namespace Characters
 				enemystats->EnemyType = ICE;
 				enemystats->health = 50 + (5 * enemystats->EnemyLevel);
 				enemystats->maxhealth = 50 + (5 * enemystats->EnemyLevel);
-				enemystats->enemytypedamage = 15 + (2 * enemystats->EnemyLevel);
+				enemystats->enemytypedamage = 15.0f + (f32)(2 * enemystats->EnemyLevel);
 				enemystats->damage = enemystats->enemytypedamage;
 				enemystats->EnemyCD = 2.0f;				//Cooldown till next enemy attack
 				enemystats->DebuffCounter = 0;
@@ -913,7 +913,7 @@ namespace Characters
 				enemystats->EnemyType = FIRE;
 				enemystats->health = 50 + (5 * enemystats->EnemyLevel);
 				enemystats->maxhealth = 50 + (5 * enemystats->EnemyLevel);
-				enemystats->enemytypedamage = 10 + (1 * enemystats->EnemyLevel);
+				enemystats->enemytypedamage = 10.0f + (f32)(1 * enemystats->EnemyLevel);
 				enemystats->damage = enemystats->enemytypedamage;
 				enemystats->EnemyCD = 2.0f;				//Cooldown till next enemy attack
 				enemystats->DebuffCounter = 0;
@@ -928,7 +928,7 @@ namespace Characters
 			enemystats->EnemyType = ENEMYBOSS;
 			enemystats->health = 70 + (5 * enemystats->EnemyLevel);
 			enemystats->maxhealth = 70 + (5 * enemystats->EnemyLevel);
-			enemystats->enemytypedamage = 20 + (5 * enemystats->EnemyLevel);
+			enemystats->enemytypedamage = 20.0f + (f32)(5 * enemystats->EnemyLevel);
 			enemystats->damage = enemystats->enemytypedamage;
 			enemystats->EnemyCD = 2.0f;				//Cooldown till next enemy attack
 			enemystats->DebuffCounter = 0;
